Fix build, update data types and improve timeout handling
Significant changes to the `Monero.Common` namespace to change the data types from `long` to `ulong` for timeouts and polling periods, improving the handling of non-negative values. Added `Subarray` function to `GenUtils.cs` to extract subarrays with proper checks. Updated methods in `MoneroConnection.cs`, `MoneroConnectionManager.cs` and `MoneroRpcConnection.cs` to use `ulong`. Modified tests in `TestMoneroConnectionManager.cs` to reflect these changes. Improved code robustness in `MoneroUtils.cs` and added new classes in `MoneroDaemonUpdateDownloadResult.cs` and `MoneroMessageSignatureResult.cs` for more flexible handling of results and signatures.

+++ Monero.Test/TestMoneroWalletCommon.cs
@@ -442,7 +442,7 @@ namespace Monero.Test
             Assert.That(false == wallet.IsConnectedToDaemon());
 
             // start polling connections
-            connectionManager.StartPolling(TestUtils.SYNC_PERIOD_IN_MS);
+            connectionManager.StartPolling((ulong)TestUtils.SYNC_PERIOD_IN_MS);
 
             // test that wallet auto connects
             Thread.Sleep(TestUtils.AUTO_CONNECT_TIMEOUT_MS);
